ref
Разбить все сервисы на админ, авторизованные, гости. перенсти в соотв. пакеты.
