+++ ewm-service/src/main/java/ru/practicum/ewm/request/service/authorized/impl/RequestAuthorizedService.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewm.request.service.impl;
+package ru.practicum.ewm.request.service.authorized.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -9,13 +9,13 @@ import ru.practicum.ewm.request.dto.RequestDtoOut;
 import ru.practicum.ewm.request.mapper.RequestMapper;
 import ru.practicum.ewm.request.model.Request;
 import ru.practicum.ewm.request.repository.RequestRepository;
-import ru.practicum.ewm.request.service.IRequestService;
+import ru.practicum.ewm.request.service.authorized.IRequestAuthorizedService;
 import ru.practicum.ewm.user.model.User;
 import ru.practicum.ewm.user.repository.UserRepository;
 
 @Service
 @RequiredArgsConstructor
-public class RequestService implements IRequestService {
+public class RequestAuthorizedService implements IRequestAuthorizedService {
     private final RequestRepository requestRepository;
     private final UserRepository userRepository;
     private final EventRepository eventRepository;
