+++ ewm-service/src/main/java/ru/practicum/ewm/user/service/admin/IUserAdminService.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewm.user.service;
+package ru.practicum.ewm.user.service.admin;
 
 import org.springframework.data.domain.Pageable;
 import ru.practicum.ewm.user.dto.UserDtoIn;
@@ -6,7 +6,7 @@ import ru.practicum.ewm.user.dto.UserDtoOut;
 
 import java.util.List;
 
-public interface IUserService {
+public interface IUserAdminService {
     UserDtoOut create(UserDtoIn userDtoIn);
 
     void delete(Long userId);
